Change mmap to take a 64 bit offset value
The mmap functions used in linux amd64 currently takes a 32 bit offset
value. This prevents passing in offsets greater than 4GB. Hence,
changing the offset value to be a uintptr which is a 64bit value.
Also changed the datatype of offset to uintptr from int in file mapping
functions.

+++ src/runtime/cgo/gcc_mmap.c
@@ -12,7 +12,7 @@
 #include "libcgo.h"
 
 uintptr_t
-x_cgo_mmap(void *addr, uintptr_t length, int32_t prot, int32_t flags, int32_t fd, uint32_t offset) {
+x_cgo_mmap(void *addr, uintptr_t length, int32_t prot, int32_t flags, int32_t fd, uintptr_t offset) {
 	void *p;
 
 	_cgo_tsan_acquire();
